# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gym Multi-Tenant API",
        "description": "API documentation for the Laravel Gym Management system",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/subscription-plans": {
            "get": {
                "tags": [
                    "Subscription Plans"
                ],
                "summary": "List all subscription plans",
                "operationId": "888991059abc64fc4a0c9f763bba7d89",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subscription Plans"
                ],
                "summary": "Create a new subscription plan",
                "operationId": "7d1259173d96f3c795e0468995e6faae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubscriptionPlanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/subscription-plans/{id}": {
            "get": {
                "tags": [
                    "Subscription Plans"
                ],
                "summary": "Get a specific subscription plan",
                "operationId": "2e81dd514abf026372a0b90c7afe0a96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subscription Plans"
                ],
                "summary": "Update a subscription plan",
                "operationId": "ac4d4102436fdb91e66ca413da87a901",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubscriptionPlanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subscription Plans"
                ],
                "summary": "Delete a subscription plan",
                "operationId": "752148f02adfe4ca786e8ad7cb50ba58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreSubscriptionPlanRequest": {
                "required": [
                    "name",
                    "price",
                    "duration"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Gold Plan"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    },
                    "duration": {
                        "description": "Duration in days",
                        "type": "integer",
                        "example": 30
                    },
                    "description": {
                        "type": "string",
                        "example": "Access to all gym facilities"
                    }
                },
                "type": "object"
            },
...
            "UpdateSubscriptionPlanRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Gold Plan"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    },
                    "duration": {
                        "description": "Duration in days",
                        "type": "integer",
                        "example": 30
                    },
                    "description": {
                        "type": "string",
                        "example": "Access to all gym facilities"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}